Fixes #6 - vertical scroll bar on iOS
Using window.innerHeight instead of
document.documentElement.clientHeight
for iOS devices.

We stay with the old height detection for the other browser, because it is the "correct way" to do this in our case (clientHeight would subtract the height of an existing horizontal scrollbar) and the browser support seems to be better on other browsers.